--- packages/api/src/deposit-address-api.ts
@@ -1,1 +1,0 @@
-export * from "./deposit-address-api/index";
