+++ packages/api/src/deposit-address.ts
@@ -1,0 +1,1 @@
+export * from "./deposit-address/index";
