+++ packages/deposit-address/.gitignore
@@ -1,5 +1,6 @@
 node_modules/
 build/
+helpers/
 *.d.ts
 *.js
