+++ packages/deposit-address/src/helpers/index.ts
@@ -1,4 +1,3 @@
 export * from "./getDepositAddressFromAxelarNetwork";
 export * from "./validation";
 export * from "./account";
-export * from "./utils";
