+++ packages/utils/src/index.ts
@@ -7,5 +7,6 @@ export * from "./function";
 export * from "./types";
 export * from "./case-conversion";
 export * from "./hex";
-
+export * from "./poll";
+export * from "./sleep";
 export { default as invariant } from "tiny-invariant";
