+++ packages/utils/src/sleep.ts
@@ -1,0 +1,3 @@
+export async function sleep(ms = 1000) {
+  return new Promise((res) => setTimeout(res, ms));
+}
